requirements.txt
numpy==1.11.0
Pillow==3.2.0
scipy==0.17.1
